% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emit_error.R
\name{emit_error}
\alias{emit_error}
\title{Stop function and provide associated messages}
\usage{
emit_error(..., .format = NULL, .f_name = TRUE, .issue = TRUE)
}
\arguments{
\item{...}{a collection of string expressions
and named arguments for string interpolation.}

\item{.format}{a format template for a
message. If not provided, the default template
of \code{"{.f_name}: {text}"} will be used.}

\item{.f_name}{the name of the function that
caused the error. If not provided, the
function name will be obtained from the
function call stack.}

\item{.issue}{a logical value that indicates
whether a warning should be issued at all.}
}
\description{
Stop the execution of the function in the
parent environment and use a message body that
can be formed from multiple text objects.
If a single text object is provided then the
warning message text will appear on a single
line. If multiple text fragments are provided
then they will be separated by newlines (in
the order provided). Custom formatting is
possible by providing a messaging template
that uses the string expression scheme used
in the \code{glue} package.
}
