% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging-package.R
\docType{package}
\name{messaging-package}
\alias{messaging}
\alias{messaging-package}
\title{messaging: Conveniently Issue Messages, Warnings, and Errors}
\description{
Provides tools for creating and issuing nicely-formatted
text within R diagnostic messages and those messages given during
warnings and errors. The formatting of the messages can be
customized using templating features. Issues with singular and
plural forms can be handled through specialized syntax.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rich-iannone/messaging}
  \item Report bugs at \url{https://github.com/rich-iannone/messaging/issues}
}

}
\author{
\strong{Maintainer}: Richard Iannone \email{riannone@me.com} (0000-0003-3925-190X)

}
\keyword{internal}
