% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{squash_all_text}
\alias{squash_all_text}
\title{Squash all text provided with a common separator}
\usage{
squash_all_text(..., .separator = "\\n")
}
\arguments{
\item{...}{a collection of text objects, which can
exist as a mixture of lists, vectors, or named
arguments.}

\item{.separator}{the separator used between
each of the text components.}
}
\description{
Squash all text provided with a common separator
}
